+++ src/js/stores/transforms/AppFormFieldToModelTransforms.js
@@ -78,6 +78,14 @@ const AppFormFieldToModelTransforms = {
     }, {});
   },
   instances: (value) => parseInt(value, 10),
+  labels: (rows) => {
+    return rows.reduce((memo, row) => {
+      if (!Util.isEmptyString(row.key) || !Util.isEmptyString(row.value)) {
+        memo[row.key] = row.value;
+      }
+      return memo;
+    }, {});
+  },
   mem: (value) => parseFloat(value),
   ports: (ports) => lazy(ports.split(","))
     .map((port) => parseInt(port, 10))